+++ saleor/account/models.py
@@ -4,7 +4,7 @@ from uuid import uuid4
 
 from django.conf import settings
 from django.contrib.auth.base_user import AbstractBaseUser, BaseUserManager
-from django.contrib.postgres.indexes import GinIndex
+from django.contrib.postgres.indexes import BTreeIndex, GinIndex
 from django.db import models
 from django.db.models import JSONField, Q, Value
 from django.db.models.expressions import Exists, OuterRef
@@ -232,6 +232,10 @@ class User(
                 fields=["last_name"],
                 name="last_name_gin",
                 opclasses=["gin_trgm_ops"],
+            ),
+            BTreeIndex(
+                fields=["date_joined"],
+                name="user_date_joined_idx",
             ),
         ]