+++ saleor/graphql/page/schema.py
@@ -2,14 +2,14 @@ import graphene
 
 from ..core import ResolveInfo
 from ..core.connection import create_connection_slice, filter_connection_queryset
-from ..core.descriptions import ADDED_IN_321
+from ..core.descriptions import ADDED_IN_321, ADDED_IN_322, DEPRECATED_IN_3X_INPUT
 from ..core.doc_category import DOC_CATEGORY_PAGES
 from ..core.enums import LanguageCodeEnum
 from ..core.fields import BaseField, FilterConnectionField
 from ..core.utils import from_global_id_or_error
 from ..translations.mutations import PageTranslate
 from .bulk_mutations import PageBulkDelete, PageBulkPublish, PageTypeBulkDelete
-from .filters import PageFilterInput, PageTypeFilterInput
+from .filters import PageFilterInput, PageTypeFilterInput, PageWhereInput, search_pages
 from .mutations import (
     PageAttributeAssign,
     PageAttributeUnassign,
@@ -48,7 +48,18 @@ class PageQueries(graphene.ObjectType):
     pages = FilterConnectionField(
         PageCountableConnection,
         sort_by=PageSortingInput(description="Sort pages."),
-        filter=PageFilterInput(description="Filtering options for pages."),
+        filter=PageFilterInput(
+            description=(
+                "Filtering options for pages."
+                f"{DEPRECATED_IN_3X_INPUT} + Use `where` and `search` instead."
+            ),
+        ),
+        search=graphene.String(
+            description="Search pages. Overrides filter.search input." + ADDED_IN_322
+        ),
+        where=PageWhereInput(
+            description="Where filtering options for pages." + ADDED_IN_322
+        ),
         description="List of the shop's pages.",
         doc_category=DOC_CATEGORY_PAGES,
     )
@@ -77,6 +88,9 @@ class PageQueries(graphene.ObjectType):
     @staticmethod
     def resolve_pages(_root, info: ResolveInfo, **kwargs):
         qs = resolve_pages(info)
+        search = kwargs.get("search") or kwargs.get("filter", {}).get("search")
+        if search:
+            qs = search_pages(qs, search)
         qs = filter_connection_queryset(
             qs, kwargs, allow_replica=info.context.allow_replica
         )